+++ dart/geodata/lib/src/utils/feature_future_adapter.dart
@@ -1,4 +1,4 @@
-// Copyright (c) 2020-2023 Navibyte (https://navibyte.com). All rights reserved.
+// Copyright (c) 2020-2024 Navibyte (https://navibyte.com). All rights reserved.
 // Use of this source code is governed by a “BSD-3-Clause”-style license that is
 // specified in the LICENSE file.
 //
@@ -14,14 +14,14 @@ import '/src/core/features/feature_failure.dart';
 /// Maps a JSON Object read from [source] to an entity using [toEntity].
 ///
 /// The source function returns a future that fetches data from a file, a web
-/// resource or other sources. Contents must be GeoJSON compliant data.
+/// resource or other sources. Content must be GeoJSON compliant data.
 @internal
 Future<T> readEntityFromJsonObject<T>(
   Future<String> Function() source, {
   required T Function(Map<String, dynamic> data) toEntity,
 }) async {
   try {
-    // read contents as text
+    // read content as text
     final text = await source();
 
     // decode JSON and expect a JSON Object as `Map<String, dynamic>`
@@ -29,6 +29,29 @@ Future<T> readEntityFromJsonObject<T>(
 
     // map JSON Object to an entity
     return toEntity(data);
+  } on ServiceException<FeatureFailure> {
+    rethrow;
+  } catch (e, st) {
+    // other exceptions (including errors)
+    throw ServiceException(FeatureFailure.clientError, cause: e, trace: st);
+  }
+}
+
+/// Maps text data read from [source] to an entity using [toEntity].
+///
+/// The source function returns a future that fetches data from a file, a web
+/// resource or other sources. Content must be GeoJSON compliant data.
+@internal
+Future<T> readEntityFromText<T>(
+  Future<String> Function() source, {
+  required T Function(String text) toEntity,
+}) async {
+  try {
+    // read content as text
+    final text = await source();
+
+    // map text to an entity
+    return toEntity(text);
   } on ServiceException<FeatureFailure> {
     rethrow;
   } catch (e, st) {